+++ lib/src/react_client/component_registration_utils.dart
@@ -2,7 +2,7 @@ import 'package:react/react.dart';
 import 'package:react/react_client/bridge.dart';
 import 'package:react/react_client/js_backed_map.dart';
 import 'package:react/react_client/react_interop.dart';
-import 'package:react/react_client/react_proxies.dart';
+import 'package:react/react_client/component_factory.dart';
 
 import 'package:react/src/react_client/react_statics.dart';
 import 'package:react/src/react_client/utils.dart';
